+++ app/models/carto/organization.rb
@@ -6,24 +6,60 @@ require_dependency 'carto/helpers/organization_commons'
 
 module Carto
   class Organization < ActiveRecord::Base
+
+    DEFAULT_GEOCODING_QUOTA = 0
+    DEFAULT_HERE_ISOLINES_QUOTA = 0
+    DEFAULT_OBS_SNAPSHOT_QUOTA = 0
+    DEFAULT_OBS_GENERAL_QUOTA = 0
+    DEFAULT_MAPZEN_ROUTING_QUOTA = nil
+
+    include CartodbCentralSynchronizable
     include DataServicesMetricsHelper
     include AuthTokenGenerator
-    include Carto::OrganizationSoftLimits
+    include OrganizationSoftLimits
     include Carto::OrganizationCommons
 
-    serialize :auth_saml_configuration, CartoJsonSymbolizerSerializer
-    before_validation :ensure_auth_saml_configuration
-    validates :auth_saml_configuration, carto_json_symbolizer: true
-
-    has_many :users, -> { order(:username) }, inverse_of: :organization
     belongs_to :owner, class_name: Carto::User, inverse_of: :owned_organization
+    has_many :users, -> { order(:username) }, inverse_of: :organization
     has_many :groups, -> { order(:display_name) }, inverse_of: :organization
     has_many :assets, class_name: Carto::Asset, dependent: :destroy, inverse_of: :organization
     has_many :notifications, -> { order('created_at DESC') }, dependent: :destroy
     has_many :connector_configurations, inverse_of: :organization, dependent: :destroy
     has_many :oauth_app_organizations, inverse_of: :oauth_app, dependent: :destroy
 
-    before_destroy :destroy_groups_with_extension
+    validates :quota_in_bytes, :seats, presence: true
+    validates(
+      :name,
+      presence: true,
+      uniqueness: true,
+      format: { with: /\A[a-z0-9\-]+\z/, message: 'must only contain lowercase letters, numbers & hyphens' }
+    )
+    validates(
+      :geocoding_quota,
+      :here_isolines_quota,
+      :obs_snapshot_quota,
+      :obs_general_quota,
+      numericality: { only_integer: true }
+    )
+    validates :default_quota_in_bytes, numericality: { only_integer: true, allow_nil: true, greater_than: 0 }
+    validates :auth_saml_configuration, carto_json_symbolizer: true
+    validate :validate_password_expiration_in_d
+    validate :organization_name_collision
+    validate :validate_seats
+    validate :authentication_methods_available
+
+    before_validation :ensure_auth_saml_configuration
+    before_validation :set_default_quotas
+    before_save :register_modified_quotas
+    after_save :save_metadata
+    before_destroy :destroy_related_resources
+    after_destroy :destroy_metadata
+
+    serialize :auth_saml_configuration, CartoJsonSymbolizerSerializer
+
+    ## AR compatibility
+    alias values attributes
+    ## ./ AR compatibility
 
     def self.find_by_database_name(database_name)
       Carto::Organization.
@@ -59,42 +95,70 @@ module Carto
       end
     end
 
-    def get_geocoding_calls(options = {})
-      require_organization_owner_presence!
-      date_to = (options[:to] ? options[:to].to_date : Date.today)
-      date_from = (options[:from] ? options[:from].to_date : owner.last_billing_cycle)
-      get_organization_geocoding_data(self, date_from, date_to)
+    def default_password_expiration_in_d
+      Cartodb.get_config(:passwords, 'expiration_in_d')
     end
 
-    def period_end_date
-      owner&.period_end_date
+    def valid_builder_seats?(users = [])
+      remaining_seats(excluded_users: users).positive?
+    end
+
+    def remaining_seats(excluded_users: [])
+      seats - assigned_seats(excluded_users: excluded_users)
+    end
+
+    def assigned_seats(excluded_users: [])
+      builder_users.count { |u| !excluded_users.map(&:id).include?(u.id) }
+    end
+
+    def valid_disk_quota?(quota = default_quota_in_bytes)
+      unassigned_quota >= quota
+    end
+
+    # Make code more uniform with user.database_schema
+    def database_schema
+      name
+    end
+
+    ####
+
+    def quota_dates(options)
+      date_to = (options[:to] ? options[:to].to_date : Date.today)
+      date_from = (options[:from] ? options[:from].to_date : last_billing_cycle)
+      return date_from, date_to
+    end
+
+    def last_billing_cycle
+      owner ? owner.last_billing_cycle : Date.today
+    end
+
+    def get_geocoding_calls(options = {})
+      require_organization_owner_presence!
+      date_from, date_to = quota_dates(options)
+      get_organization_geocoding_data(self, date_from, date_to)
     end
 
     def get_here_isolines_calls(options = {})
       require_organization_owner_presence!
-      date_to = (options[:to] ? options[:to].to_date : Date.today)
-      date_from = (options[:from] ? options[:from].to_date : owner.last_billing_cycle)
+      date_from, date_to = quota_dates(options)
       get_organization_here_isolines_data(self, date_from, date_to)
     end
 
     def get_mapzen_routing_calls(options = {})
       require_organization_owner_presence!
-      date_to = (options[:to] ? options[:to].to_date : Date.today)
-      date_from = (options[:from] ? options[:from].to_date : owner.last_billing_cycle)
+      date_from, date_to = quota_dates(options)
       get_organization_mapzen_routing_data(self, date_from, date_to)
     end
 
     def get_obs_snapshot_calls(options = {})
       require_organization_owner_presence!
-      date_to = (options[:to] ? options[:to].to_date : Date.today)
-      date_from = (options[:from] ? options[:from].to_date : owner.last_billing_cycle)
+      date_from, date_to = quota_dates(options)
       get_organization_obs_snapshot_data(self, date_from, date_to)
     end
 
     def get_obs_general_calls(options = {})
       require_organization_owner_presence!
-      date_to = (options[:to] ? options[:to].to_date : Date.today)
-      date_from = (options[:from] ? options[:from].to_date : owner.last_billing_cycle)
+      date_from, date_to = quota_dates(options)
       get_organization_obs_general_data(self, date_from, date_to)
     end
 
@@ -135,6 +199,70 @@ module Carto
       (remaining > 0 ? remaining : 0)
     end
 
+    def to_poro
+      {
+        created_at:       created_at,
+        description:      description,
+        discus_shortname: discus_shortname,
+        display_name:     display_name,
+        id:               id,
+        name:             name,
+        owner: {
+          id:         owner ? owner.id : nil,
+          username:   owner ? owner.username : nil,
+          avatar_url: owner ? owner.avatar_url : nil,
+          email:      owner ? owner.email : nil,
+          groups:     owner && owner.groups ? owner.groups.map { |g| Carto::Api::GroupPresenter.new(g).to_poro } : []
+        },
+        admins:                    users.select(&:org_admin).map { |u| { id: u.id } },
+        quota_in_bytes:            quota_in_bytes,
+        unassigned_quota:          unassigned_quota,
+        geocoding_quota:           geocoding_quota,
+        map_view_quota:            map_view_quota,
+        twitter_datasource_quota:  twitter_datasource_quota,
+        map_view_block_price:      map_view_block_price,
+        geocoding_block_price:     geocoding_block_price,
+        here_isolines_quota:       here_isolines_quota,
+        here_isolines_block_price: here_isolines_block_price,
+        obs_snapshot_quota:        obs_snapshot_quota,
+        obs_snapshot_block_price:  obs_snapshot_block_price,
+        obs_general_quota:         obs_general_quota,
+        obs_general_block_price:   obs_general_block_price,
+        geocoder_provider:         geocoder_provider,
+        isolines_provider:         isolines_provider,
+        routing_provider:          routing_provider,
+        mapzen_routing_quota:       mapzen_routing_quota,
+        mapzen_routing_block_price: mapzen_routing_block_price,
+        seats:                     seats,
+        twitter_username:          twitter_username,
+        location:                  twitter_username,
+        updated_at:                updated_at,
+        website:                   website,
+        admin_email:               admin_email,
+        avatar_url:                avatar_url,
+        user_count:                users.count,
+        password_expiration_in_d:  password_expiration_in_d
+      }
+    end
+
+    def tags(type, exclude_shared=true)
+      users.map { |u| u.tags(exclude_shared, type) }.flatten
+    end
+
+    def public_vis_by_type(type, page_num, items_per_page, tags, order = 'updated_at', version = nil)
+      CartoDB::Visualization::Collection.new.fetch(
+          user_id:  self.users.map(&:id),
+          type:     type,
+          privacy:  CartoDB::Visualization::Member::PRIVACY_PUBLIC,
+          page:     page_num,
+          per_page: items_per_page,
+          tags:     tags,
+          order:    order,
+          o:        { updated_at: :desc },
+          version:  version
+      )
+    end
+
     def signup_page_enabled
       whitelisted_email_domains.present? && auth_enabled?
     end
@@ -143,8 +271,32 @@ module Carto
       auth_username_password_enabled || auth_google_enabled || auth_github_enabled || auth_saml_enabled?
     end
 
+    def total_seats
+      seats + viewer_seats
+    end
+
+    def remaining_viewer_seats(excluded_users: [])
+      viewer_seats - assigned_viewer_seats(excluded_users: excluded_users)
+    end
+
+    def assigned_viewer_seats(excluded_users: [])
+      viewer_users.count { |u| !excluded_users.map(&:id).include?(u.id) }
+    end
+
+    def notify_if_disk_quota_limit_reached
+      ::Resque.enqueue(::Resque::OrganizationJobs::Mail::DiskQuotaLimitReached, id) if disk_quota_limit_reached?
+    end
+
+    def notify_if_seat_limit_reached
+      ::Resque.enqueue(::Resque::OrganizationJobs::Mail::SeatLimitReached, id) if seat_limit_reached?
+    end
+
     def database_name
-      owner ? owner.database_name : nil
+      owner&.database_name
+    end
+
+    def revoke_cdb_conf_access
+      users.map { |user| user.db_service.revoke_cdb_conf_access }
     end
 
     def create_group(display_name)
@@ -152,15 +304,31 @@ module Carto
     end
 
     def name_to_display
-      display_name.nil? ? name : display_name
+      display_name || name
+    end
+
+    def max_import_file_size
+      owner ? owner.max_import_file_size : ::User::DEFAULT_MAX_IMPORT_FILE_SIZE
+    end
+
+    def max_import_table_row_count
+      owner ? owner.max_import_table_row_count : ::User::DEFAULT_MAX_IMPORT_TABLE_ROW_COUNT
+    end
+
+    def max_concurrent_import_count
+      owner ? owner.max_concurrent_import_count : ::User::DEFAULT_MAX_CONCURRENT_IMPORT_COUNT
+    end
+
+    def max_layers
+      owner ? owner.max_layers : ::User::DEFAULT_MAX_LAYERS
     end
 
     def assigned_quota
-      self.users.sum(:quota_in_bytes).to_i
+      users.sum(:quota_in_bytes).to_i
     end
 
     def unassigned_quota
-      self.quota_in_bytes - assigned_quota
+      quota_in_bytes - assigned_quota
     end
 
     def require_organization_owner_presence!
@@ -186,7 +354,7 @@ module Carto
     end
 
     def non_owner_users
-      users.reject { |u| owner && u.id == owner.id }
+      owner ? users.where.not(id: owner.id) : users
     end
 
     def inheritable_feature_flags
@@ -210,18 +378,121 @@ module Carto
       users.map { |u| u.get_api_calls(options).sum }.sum
     end
 
+    def require_organization_owner_presence!
+      raise Carto::Organization::OrganizationWithoutOwner.new(self) unless owner
+    end
+
+    ## TODO: make private once model is fully migrated
+
+    def destroy_non_owner_users
+      non_owner_users.each do |user|
+        user.ensure_nonviewer
+        user.shared_entities.map(&:entity).uniq.each(&:delete)
+        user.sequel_user.destroy_cascade
+      end
+    end
+
+    def destroy_assets
+      assets.map { |asset| Carto::Asset.find(asset.id) }.map(&:destroy).all?
+    end
+
+    # Returns true if disk quota won't allow new signups with existing defaults
+    def disk_quota_limit_reached?
+      unassigned_quota < default_quota_in_bytes
+    end
+
+    # Returns true if seat limit will be reached with new user
+    def seat_limit_reached?
+      (remaining_seats - 1) < 1
+    end
+
+    def public_vis_count_by_type(type)
+      CartoDB::Visualization::Collection.new.fetch(
+        user_id: users.map(&:id),
+        type: type,
+        privacy: CartoDB::Visualization::Member::PRIVACY_PUBLIC,
+        per_page: CartoDB::Visualization::Collection::ALL_RECORDS
+      ).count
+    end
+
+    # INFO: replacement for destroy because destroying owner triggers
+    # organization destroy
+    def destroy_cascade(delete_in_central: false)
+      # This remains commented because we consider that enabling this for users at SaaS is unnecessary and risky.
+      # Nevertheless, code remains, _just in case_. More info at https://github.com/CartoDB/cartodb/issues/12049
+      # Central branch: 1764-Allow_updating_inactive_users
+      # Central asks for usage information before deleting, so organization must be first deleted there
+      # Corollary: you need multithreading for organization to work if you run Central
+      # self.delete_in_central if delete_in_central
+
+      groups.each(&:destroy_group_with_extension)
+      destroy_non_owner_users
+      owner ? owner.sequel_user.destroy_cascade : destroy
+    end
+
+    def validate_seats_for_signup(user, errors)
+      errors.add(:organization, 'not enough seats') if user.builder? && !valid_builder_seats?([user])
+
+      return unless user.viewer? && remaining_viewer_seats(excluded_users: [user]) <= 0
+
+      errors.add(:organization, 'not enough viewer seats')
+    end
+
+    def validate_for_signup(errors, user)
+      validate_seats_for_signup(user, errors)
+
+      return if valid_disk_quota?(user.quota_in_bytes.to_i)
+
+      errors.add(:quota_in_bytes, 'not enough disk quota')
+    end
+
     private
 
+    def authentication_methods_available
+      if whitelisted_email_domains.present? && !auth_enabled?
+        errors.add(:whitelisted_email_domains, 'enable at least one auth. system or clear whitelisted email domains')
+      end
+    end
+
+    def organization_name_collision
+      errors.add(:name, 'cannot exist as user') if Carto::User.exists?(username: name)
+    end
+
+    def validate_seats
+      errors.add(:seats, 'cannot be less than the number of builders') if seats && remaining_seats.negative?
+
+      return unless viewer_seats && remaining_viewer_seats.negative?
+
+      errors.add(:viewer_seats, 'cannot be less than the number of viewers')
+    end
+
+    def validate_password_expiration_in_d
+      valid = password_expiration_in_d.blank? || password_expiration_in_d.positive? && password_expiration_in_d < 366
+      errors.add(:password_expiration_in_d, 'must be greater than 0 and lower than 366') unless valid
+    end
+
     def ensure_auth_saml_configuration
-      self.auth_saml_configuration ||= Hash.new
+      self.auth_saml_configuration ||= {}
     end
 
-    def destroy_groups_with_extension
-      return unless groups
-
-      groups.map { |g| g.destroy_group_with_extension }
-
-      reload
+    def set_default_quotas
+      self.geocoding_quota ||= DEFAULT_GEOCODING_QUOTA
+      self.here_isolines_quota ||= DEFAULT_HERE_ISOLINES_QUOTA
+      self.obs_snapshot_quota ||= DEFAULT_OBS_SNAPSHOT_QUOTA
+      self.obs_general_quota ||= DEFAULT_OBS_GENERAL_QUOTA
+      self.mapzen_routing_quota ||= DEFAULT_MAPZEN_ROUTING_QUOTA
     end
+
+    # TODO: These variables are not read explicitly. Can they be removed?
+    def register_modified_quotas
+      @geocoding_quota_modified = geocoding_quota_changed?
+      @here_isolines_quota_modified = here_isolines_quota_changed?
+      @obs_snapshot_quota_modified = obs_snapshot_quota_changed?
+      @obs_general_quota_modified = obs_general_quota_changed?
+      @mapzen_routing_quota_modified = mapzen_routing_quota_changed?
+
+      raise errors.join('; ') unless valid?
+    end
+
   end
 end