+++ spec/factories/organizations.rb
@@ -5,9 +5,7 @@ include UniqueNamesHelper
 
 FactoryGirl.define do
 
-  factory :organization do
-    to_create(&:save)
-
+  factory :organization, class: Carto::Organization do
     name { unique_name('organization') }
     seats 10
     quota_in_bytes 100.megabytes
@@ -27,17 +25,17 @@ FactoryGirl.define do
     isolines_provider 'heremaps'
     routing_provider 'heremaps'
 
-    factory :organization_whitelist_carto do
+    factory :organization_whitelist_carto, class: Carto::Organization do
       whitelisted_email_domains ['carto.com']
       auth_username_password_enabled true
     end
 
-    factory :organization_google_whitelist_empty do
+    factory :organization_google_whitelist_empty, class: Carto::Organization do
       whitelisted_email_domains []
       auth_google_enabled true
     end
 
-    factory :organization_with_users do
+    factory :organization_with_users, class: Carto::Organization do
       after(:create) do |org|
         create_account_type_fg('ORGANIZATION USER')
         owner = FactoryGirl.create(:user)
@@ -63,7 +61,7 @@ FactoryGirl.define do
       end
     end
 
-    factory :saml_organization do
+    factory :saml_organization, class: Carto::Organization do
       auth_saml_configuration do
         {
           issuer: 'localhost.lan',
@@ -78,8 +76,6 @@ FactoryGirl.define do
     end
 
     factory :carto_organization, class: Carto::Organization do
-      to_create(&:save!)
-
       name { unique_name('organization') }
       seats 10
       quota_in_bytes 100.megabytes